upsample, limit faces to subset of indices, reorder vertex and face list so face'd indices are first, all pairs distances
Former-commit-id: 1e9a117fb46ae79d0b467845558bd0f458f951dd

+++ ismanifold.h
@@ -32,7 +32,7 @@ namespace igl
       }
     std::sort(TTT.begin(),TTT.end());
     
-    for(int i=2;i<TTT.size();++i)
+    for(int i=2;i<(int)TTT.size();++i)
     {
       std::vector<int>& r1 = TTT[i-2];
       std::vector<int>& r2 = TTT[i-1];
